refactor/App: Created Home.jsx, moved main contents there, reframed routes inside app and added about route as well
